Merge pull request #120 from vjpalodichuk/develop
:ambulance: Added AOC 2017 Day 18

+++ aoc-2016/src/main/kotlin/com/capital7software/aoc/aoc2016aoc/days/Day23.kt
@@ -135,8 +135,8 @@ class Day23 : AdventOfCodeSolution {
   @SuppressFBWarnings
   fun getValueInRegister(input: List<String>, read: String, write: String, eggs: Int): Int {
     val computer = SmallComputer.buildSmallComputer(input)
-    computer[write] = eggs
+    computer[write] = eggs.toLong()
     computer.run()
-    return computer[read]
+    return computer[read].toInt()
   }
 }
